# formulae_dependents: apply some optimisations